# tooling/sparta/packages/utils/openapi/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sparta API",
    "version": "1.0.0",
    "description": "…"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Sparta API"
    }
  ],
  "components": {
    "schemas": {
...
      "NodeOperator": {
        "type": "object",
        "properties": {
          "discordId": {
            "type": "string",
            "description": "The Discord user ID of the node operator.",
            "example": "123456789012345678"
          },
          "walletAddress": {
            "type": "string",
            "description": "The Ethereum wallet address associated with the node operator.",
            "example": "0x1234567890abcdef1234567890abcdef12345678"
          },
          "createdAt": {
            "type": "number",
            "description": "Timestamp (milliseconds since epoch) when the operator was created.",
            "example": 1678886400000
          },
          "updatedAt": {
            "type": "number",
            "description": "Timestamp (milliseconds since epoch) when the operator was last updated.",
            "example": 1678887400000
          }
        },
        "required": [
          "discordId",
          "walletAddress",
          "createdAt",
          "updatedAt"
        ]
      },
      "OperatorInput": {
        "type": "object",
        "properties": {
          "discordId": {
            "type": "string",
            "description": "The Discord user ID."
          },
          "walletAddress": {
            "type": "string",
            "description": "The Ethereum wallet address.",
            "example": "0x1234567890abcdef1234567890abcdef12345678"
          }
        },
        "required": [
          "discordId",
          "walletAddress"
        ]
      },
      "OperatorUpdateInput": {
        "type": "object",
        "properties": {
          "walletAddress": {
            "type": "string",
            "description": "The new Ethereum wallet address.",
            "example": "0xabcdef1234567890abcdef1234567890abcdef12"
          }
        },
        "required": [
          "walletAddress"
        ]
      },
      "OperatorError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message describing the issue."
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/operator": {
      "get": {
        "summary": "Get all node operators",
        "description": "Retrieves a list of all registered node operators.",
        "tags": [
          "NodeOperator"
        ],
        "operationId": "getAllOperators",
        "responses": {
          "200": {
            "description": "A list of node operators.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NodeOperator"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new node operator",
        "description": "Registers a new node operator with their Discord ID and wallet address.",
        "tags": [
          "NodeOperator"
        ],
        "operationId": "createOperator",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Node operator created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeOperator"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing or invalid body parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Operator with this Discord ID already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          }
        }
      }
    },
    "/api/operator/discord/{discordId}": {
      "get": {
        "summary": "Get operator by Discord ID",
        "description": "Retrieves a specific node operator using their Discord ID.",
        "tags": [
          "NodeOperator"
        ],
        "operationId": "getOperatorByDiscordId",
        "parameters": [
          {
            "in": "path",
            "name": "discordId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The Discord ID of the operator to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "The requested node operator.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeOperator"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing discordId parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          },
          "404": {
            "description": "Operator not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          }
        }
      }
    },
    "/api/operator/address/{address}": {
      "get": {
        "summary": "Get operator by wallet address",
        "description": "Retrieves a specific node operator using their wallet address.",
        "tags": [
          "NodeOperator"
        ],
        "operationId": "getOperatorByAddress",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The wallet address of the operator to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "The requested node operator.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeOperator"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing or invalid address parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          },
          "404": {
            "description": "Operator not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          }
        }
      }
    },
    "/api/operator/{discordId}": {
      "delete": {
        "summary": "Delete an operator by Discord ID",
        "description": "Removes a node operator registration using their Discord ID.",
        "tags": [
          "NodeOperator"
        ],
        "operationId": "deleteOperatorByDiscordId",
        "parameters": [
          {
            "in": "path",
            "name": "discordId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The Discord ID of the operator to delete."
          }
        ],
        "responses": {
          "204": {
            "description": "Operator deleted successfully (No Content)."
          },
          "400": {
            "description": "Bad Request - Missing discordId parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          },
          "404": {
            "description": "Operator not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update operator's wallet address",
        "description": "Updates the wallet address for a specific node operator using their Discord ID.",
        "tags": [
          "NodeOperator"
        ],
        "operationId": "updateOperatorWallet",
        "parameters": [
          {
            "in": "path",
            "name": "discordId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The Discord ID of the operator to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorUpdateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operator updated successfully. Returns the updated operator.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeOperator"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing discordId parameter or invalid/missing walletAddress in body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          },
          "404": {
            "description": "Operator not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}